+++ src/Paprika/Globals.cs
@@ -1,4 +1,3 @@
 ﻿using System.Runtime.CompilerServices;
 
-[module: SkipLocalsInit]
 [assembly: InternalsVisibleTo("Paprika.Tests")]
