+++ src/Paprika/Pages/AccountFrame.cs
@@ -1,5 +1,6 @@
 ﻿using System.Runtime.InteropServices;
 using Nethermind.Int256;
+using Paprika.Crypto;
 
 namespace Paprika.Pages;
