+++ src/Paprika/Pages/RootPage.cs
@@ -1,5 +1,6 @@
 ﻿using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
+using Paprika.Crypto;
 
 namespace Paprika.Pages;
