+++ src/Paprika/Pages/SetContext.cs
@@ -1,4 +1,5 @@
 ﻿using Nethermind.Int256;
+using Paprika.Crypto;
 
 namespace Paprika.Pages;
