--- src/Tree.Runner/Tree.Runner.csproj
@@ -1,24 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <OutputType>Exe</OutputType>
-        <TargetFramework>net7.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-        <PlatformTarget>x64</PlatformTarget>
-        <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\Tree.Tests\Tree.Tests.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <PackageReference Include="Spectre.Console" Version="0.45.1-preview.0.43" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <Folder Include="db" />
-    </ItemGroup>
-
-</Project>
