+++ src/Tree.sln
@@ -1,13 +1,5 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tree", "Tree\Tree.csproj", "{9FE359B5-FAE1-48B8-8AE9-DE63137F721E}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tree.Tests", "Tree.Tests\Tree.Tests.csproj", "{74D5D451-4C57-4778-94DB-9691D897F59A}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tree.Runner", "Tree.Runner\Tree.Runner.csproj", "{756973B2-BF89-4238-A7E4-2894C21FD676}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tree.Crypto", "Tree.Crypto\Tree.Crypto.csproj", "{D5CE5B9A-2A76-4429-AECA-75529B9463AC}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Paprika", "Paprika\Paprika.csproj", "{18F81388-685F-4FFC-A5D8-ABB79FE254DD}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Paprika.Tests", "Paprika.Tests\Paprika.Tests.csproj", "{44B0A996-A65A-4BFE-B9F6-708D5A7E54C5}"
@@ -20,22 +12,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{9FE359B5-FAE1-48B8-8AE9-DE63137F721E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9FE359B5-FAE1-48B8-8AE9-DE63137F721E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9FE359B5-FAE1-48B8-8AE9-DE63137F721E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9FE359B5-FAE1-48B8-8AE9-DE63137F721E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{74D5D451-4C57-4778-94DB-9691D897F59A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{74D5D451-4C57-4778-94DB-9691D897F59A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{74D5D451-4C57-4778-94DB-9691D897F59A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{74D5D451-4C57-4778-94DB-9691D897F59A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{756973B2-BF89-4238-A7E4-2894C21FD676}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{756973B2-BF89-4238-A7E4-2894C21FD676}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{756973B2-BF89-4238-A7E4-2894C21FD676}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{756973B2-BF89-4238-A7E4-2894C21FD676}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D5CE5B9A-2A76-4429-AECA-75529B9463AC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D5CE5B9A-2A76-4429-AECA-75529B9463AC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D5CE5B9A-2A76-4429-AECA-75529B9463AC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D5CE5B9A-2A76-4429-AECA-75529B9463AC}.Release|Any CPU.Build.0 = Release|Any CPU
 		{18F81388-685F-4FFC-A5D8-ABB79FE254DD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{18F81388-685F-4FFC-A5D8-ABB79FE254DD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{18F81388-685F-4FFC-A5D8-ABB79FE254DD}.Release|Any CPU.ActiveCfg = Release|Any CPU
