--- src/Tree/Globals.cs
@@ -1,4 +1,0 @@
-﻿using System.Runtime.CompilerServices;
-
-[module: SkipLocalsInit]
-[assembly: InternalsVisibleTo("Tree.Tests")]
