--- src/Tree/Tree.csproj
@@ -1,14 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net7.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-        <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\Tree.Crypto\Tree.Crypto.csproj" />
-    </ItemGroup>
-
-</Project>
